// src/app/models/elementId.enum.ts
export enum elementId {
  appVersionOutput = 'app-version',
  convertButton = 'convert-btn',
  svgInput = 'svg-in',
  svgPreviewContainer = 'svg-preview-container',
  svgPreviewInjectionEl = 'svg-preview-injection',
  previewBtn = 'svg-preview-btn',
  previewBtnIcon = 'svg-preview-btn-icon',
  previewShowingCls = 'preview-showing',
  checkboxIncludeFull = 'include-full',
  stencilOutput = 'mx-out',
}